Feature/ws tls (#16)
* checkin with changes to main

I have a node_listener object yet to check in.  Still making some adjustments to it.

* getting close to fully functional.

What is left is to spin up 1 or more bootnodes and then start connecting other nodes.

* working on discovery. Check in for other machines.

* update gitignore

* adding command line argument for testing

* connect peers for hello

* baseline startup with docker container and bootNode

* add some temporary comms to keep connections alive

* adding tools and update oracle node announce

generated code that will be changing soon.

* update gitignore

* functional tests for calls to smart contracts

Calls are working. Had to convert key type to conform to supported ECDSA for ETH integration

* update to use quic protocol

---------

Co-authored-by: Bob Stevens <35038919+restevens402@users.noreply.github.com>