working with issues with Gossip (#37)
* This has the foundation for gossiping the network participants and adding them to a ledger

I still need to get the storage going and make sure the gossip works as expected, but this includes the foundation to do all of that. The remaining work should be making sure the storage works like I mean it to.

* working on issues with gossip

something isn't working right. checking in to refresh other nodes to test

* check in for integration testing

* date check

* fix the assignment of data object that is gossiped

* trying to force adding to the peerstore

Gossip is just not working like it should

* last check in before merge