Feature/multiple multi address (#68)
* Update README.md

* Update README.md

* refactoring the pubsub for NodeData

* starting support for using array of multiaddress

* support for list of multi-addrs for a peer in the cache.

* small naming updates

* add synchronization

reduce some logging noise.

* adding more synch logging for testing

* wrapping multiAddr so that it can be serialized to JSON

* working on the public key support

* refactoring for AdHandler and startup

I moved the staking logic out of the cli init() function. It had some duplication for the private key, so I make the call to the handleStaking right after the private key is retrieved. If staking happens, I ensure that the system exits and doesn't start the node.

---------

Co-authored-by: Brendan Playford <34052452+teslashibe@users.noreply.github.com>