Refactor node event tracker for disconnected nodes
The logic for handling un-staked and non-existing nodes in the "Disconnected" function of the node event tracker was updated. Now, the system only returns when a node is not staked if the node data exists. Also, removed the usage of 'pubKeyHex', as it is not used in the updated logic.