02 07 testnet bug fixes (#126)
* Refactor node event tracker for disconnected nodes

The logic for handling un-staked and non-existing nodes in the "Disconnected" function of the node event tracker was updated. Now, the system only returns when a node is not staked if the node data exists. Also, removed the usage of 'pubKeyHex', as it is not used in the updated logic.

* Set default communication protocol and RPC URL

Added a default communication protocol (UDP) in cli.go if neither UDP nor TCP are specified by the user. Also, appended a default RPC URL (https://ethereum-sepolia.publicnode.com) to the constants module and wrote it to the environment file in main.go. These changes ensure user friendliness and default values when certain parameters are not specified.