Issue 131 (#132)
* dev -> test (#130)

* Update README.md (#124)

* Update README.md

* add link to DOCKER.md

* link to DOCKER.md

* Update README to descriptions

---------

Co-authored-by: Bob Stevens <35038919+restevens402@users.noreply.github.com>

* 02 07 testnet bug fixes (#126)

* Refactor node event tracker for disconnected nodes

The logic for handling un-staked and non-existing nodes in the "Disconnected" function of the node event tracker was updated. Now, the system only returns when a node is not staked if the node data exists. Also, removed the usage of 'pubKeyHex', as it is not used in the updated logic.

* Set default communication protocol and RPC URL

Added a default communication protocol (UDP) in cli.go if neither UDP nor TCP are specified by the user. Also, appended a default RPC URL (https://ethereum-sepolia.publicnode.com) to the constants module and wrote it to the environment file in main.go. These changes ensure user friendliness and default values when certain parameters are not specified.

* copy update for README (#125)

* update copy on README

* script to build and install masa-node as a service

* testing out impact of continuous gossip (#129)

---------

Co-authored-by: JD <156010594+5u6r054@users.noreply.github.com>

* Add environment flag to CLI and support different protocol versions

A new CLI flag 'env' is introduced to specify the environment to connect to. It makes sure the protocol version and communication topic are based on the particular environment, allowing the application to switch more readily between different environments. Also, error handling for setting environment variables has been improved.

* Fixed variable for environment flag in CLI

The "env" variable is assigned its own CLI flag, correcting a previous error where it was assigned to the "stakeAmount" variable. them.

* increment version

* Update README.md

add flag --env=test to segment testnet comms

* update DOCKER.md to tell users to set the ENV=test environment variable.

* update env vars info

* update Dockerfile and docker-compose.yaml to use the ENV setting (which helps keep nodes joining the correct network)

---------

Co-authored-by: JD <156010594+5u6r054@users.noreply.github.com>