Feature/gateway (#249)
* Added LLM Chat support and route

Added LlmChatUrl constant to the config package and an LLMChat() function to the API. This function reads the incoming request, proxies it to the configured LLM Chat URL, and returns the response. Also added a new POST route "/chat" for LLM chat functionality.

* update dependencies

* Add LLM Chat feature and endpoint in Masa Node

Implemented Large Language Model (LLM) Chat feature in Masa Node that allows users to interact with multiple models through a specific endpoint. The functionality acts as a proxy that forwards user requests to the respective LLMs. Detailed documentation about LLM Chat usage and response structure is provided.

* Update dependencies versions in go.mod and go.sum

The main dependency versions have been updated in go.mod. This also includes the corresponding changes in go.sum file that reflect the new versions. These changes will help ensure the project uses the latest and most secure versions of each dependency.

* merge

* using LLM_CHAT_URL

* added pow PoC for api bearer token for gateway

* added --version for command line and version in welcome

* added JWT option for gateway to demonstrate differences

* wip auth endpooint to generate api keys for nodes

* added mock llm0-chat actor

* removed LLM sentiment prompt from env since to ensure masa manages the expected outcome, we will want to pull this from an administrated database

* updating release notes

* updated version 0.0.3-x

* moved add subscription to monitor worker fn

* changed all command line params to camelCase

* using config instead of getting from os env.

* Add chat functionality to worker services

This commit implements chat functionality into the worker services by publishing messages directed towards the chat url on the worker topic. The responses are then processed and returned. A new `client.go` file is also included to facilitate HTTP requests within the worker services. Other minor changes include tweaks to importations and processing of username data for login functionality.

* updated visit links for maxdepth bugfix

* md lint docs

* Refactor actor participation checks in workers package

An IsActor method was added to the OracleNode struct to centralize the participation checks in the workers file. This method checks if the node is a WebScraper, a TwitterScraper, or a LlmServer.

* updated visit depth pages

* using test pg for storing raw data until gateway is ready, adding cf ai workers for PoC

* testing cf dev config

* using cf for llm chat PoC

* updated dockerfile with restart always, added /chat/cf to test cloudflare as ai workers server

* cycled cf token

* finished pkgs for scrapers and wip oracle work data

* use masa-twitter-scraper

* added authorization login for swagger use /auth endpoint to generate api key for your node

* updated /chat for llm playground - requires local ollama atm

* Implement request-response pattern in workers

This commit introduces a change in the worker's logic to use a request-response pattern. It establishes a unique request identifier and uses a newly created ResponseChannelMap to manage response channels for each request. Mappings are cleaned up after each use to prevent memory leaks.

* Update package versions in go.mod and go.sum

Updated the versions of multiple dependencies in the go.mod and go.sum files. This includes updates to toolchain, bubbletea, go-ethereum, color, ollama, and go-openai packages.

* completed added LLMs for ollama/ and @cf (cloudflare) and @hf (huggingface) support

* chat llm page /chat poc

* Added returning a response from distributed work. The Send function is replaced with a RequestFuture and in the receive function the worker returns the results to the calling node.

* Add error handling in API calls

Updated handlers_data.go to better handle errors during API calls. If there are errors during the Publish function call or during JSON unmarshalling, the function will now return an HTTP Status Expectation Failed error and the function will terminate.

* getting a little bit hacky with getting this to work.

The protobuf will have to be adjusted. It will make sense to add a new field to the message. Something like channelID

* Revert "getting a little bit hacky with getting this to work."

This reverts commit 7df159249afcc17bb589f2495e725eb62856fecc.

* getting around protobuf json serialization for now

* allow strings as validator data.

* Refactor subscription process to include self-generated messages

Updated subscription method in the pubsub manager to add an option to include or skip messages generated by the same node. Several changes to handler instantiation across the codebase to incorporate this new option. This provides more control and flexibility in handling pub-sub scenarios where same-node message processing might be needed. This is to support nodes participating in the actor pattern for distributed requests.

* adding oracle data funcs

* followers are returning on userid

* just need to tie up a few loose ends

* updated llm completion docs and updated get llm service

* add endpoint to swagger docs

* updated all dependabot issues

* updated proto buffer and workers

* updated /chat llm page to use the best performing cf ai model

* added private network allowed for cors for api and chat

* db testing

* db tests

* data funcs

* unmarshal directly into pubsub.Message

* This commit introduces updates in how worker responses are handled. Previously, worker responses were directly unmarshalled; now, they are first cast to a map, which was necessary to get past unmarshal errors from pubsub.Message. A new getResponseMessage function is also introduced for reuse in unmarshalling worker responses.

* wip actor process

* updated scraper lib without console logging

* all nodes must subscribe but only staked actor nodes monitor

* doc updates

* updated release docs

* added note inline

* documented func

* manuallly convert message when it fails on unmarshal

---------

Co-authored-by: Bob Stevens <35038919+restevens402@users.noreply.github.com>
Co-authored-by: Nolan Jacobson <50815660+nolanjacobson@users.noreply.github.com>